Fix: 꼬리질문 반환시 질문 인덱스 float 타입으로 변경

+++ app/schemas/response/interview_question_response.py
@@ -6,7 +6,7 @@ class InterviewQuestionResponse(BaseModel):
     question_id: str
     type: str
     total_questions: int
-    question_index: int
+    question_index: float
     question_text: str
     has_follow_up: bool
